+++ lawnchair/src/app/lawnchair/qsb/providers/AppSearch.kt
@@ -1,0 +1,13 @@
+package app.lawnchair.qsb.providers
+
+import app.lawnchair.qsb.ThemingMethod
+import com.android.launcher3.R
+
+object AppSearch : QsbSearchProvider(
+    id = "app-search",
+    name = "App Search",
+    icon = R.drawable.ic_qsb_search,
+    themingMethod = ThemingMethod.TINT,
+    packageName = "",
+    website = ""
+)
